First running setup with docker-compose
A complete configuration has been setup with docker-compose, it's
launching a mysql container along with an nginx one hosting version
2.1.14 of bolt CMS.

Ideally database backups should be performed using another docker
container that is going to be bound to the `/var/lib/mysql` volume on
the `bolt_db` container, archive all of it's content and put it on the
host machine. Alternatively the backup container could also upload
the backup directly to the desired location (S3, an FTP server...).