Add a volume to access themes' config files
For some themes (all the fancy HTML5-UP ones actually), we need to
modify theme-specific config files that can't be modified from
`/app/config/extensions`. Therefore the `/theme` folder has been added
as a volume (`/themes` seems to be unused) so that the themes and their
configuration can be easily backed up as well.

This also makes it possible to modify the theme's config from the docker
host or from another container mounting the volume.

Ideally, it would be great to build a docker image on top of bolt server
to embed both cronjobs for backups and to preconfigure what won't really
change once our CMS is first setup: the choice of a theme, it's config,
some static strings in that config...