Volume management refactoring
The config files for bolt come prestored in our volume (just the
`*.yml.dist` ones). The idea is that confeig, themes and MySQL data are
mounted from the host into the container.

On a local development environmenent managed by `docker-compose`, the
local `./volumes` folder is used by default but in production, different
strategy can be used (from just storing these on the single docker host
and becking them up to a file storage system at regular intervals to
using Kubernetes to have them stored into a centralized place and
mounted on each container).