We have a ready to ship image of Anchor !!
Let's roll! Now one just has to fork this repository to get started with
a blog project. The docker images doesn't have to be changed, actually
it can (and should) be pulled from Dockerhub **as is**.

The database folder should be backed up as often as possible. We'll
include a script for that.